Creating a blueprint for admin
